+++ lib/core/src/system/gpu/data/texture/types.rs
@@ -1,0 +1,15 @@
+//! This is a root module for all texture-related data types.
+
+pub mod format;
+pub mod gl_enums;
+pub mod internal_format;
+pub mod item;
+pub mod relations;
+pub mod sampler;
+
+pub use format::*;
+pub use gl_enums::*;
+pub use internal_format::*;
+pub use item::*;
+pub use relations::*;
+pub use sampler::*;
