+++ lib/core/src/system/gpu/data/texture/types/gl_enums.rs
