+++ lib/core/src/system/gpu/data/texture/types/internal_format.rs
