+++ lib/core/src/system/gpu/data/texture/types/relations.rs
