+++ lib/core/src/system/gpu/data/texture/types/sampler.rs
